refactor: Move cache service from Service Infra to Persistence Infra
