+++ Kathanika.sln
@@ -13,8 +13,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Kathanika.Infrastructure.Gr
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Kathanika.Infrastructure.Persistence", "src\Infrastructure.Persistence\Kathanika.Infrastructure.Persistence.csproj", "{EF9F227B-52C2-4CF0-BB74-C5597BD5F627}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Kathanika.Infrastructure.Services", "src\Infrastructure.Services\Kathanika.Infrastructure.Services.csproj", "{784EC1C5-0607-4B9B-A53A-B2636CE756A6}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Kathanika.Web", "src\Web\Kathanika.Web.csproj", "{F12B6A05-AB25-4D95-85FD-FA1BD7164FE2}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{0B38BBD6-2FA3-4AAE-960B-B264155C6802}"
@@ -45,10 +43,6 @@ Global
 		{EF9F227B-52C2-4CF0-BB74-C5597BD5F627}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{EF9F227B-52C2-4CF0-BB74-C5597BD5F627}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{EF9F227B-52C2-4CF0-BB74-C5597BD5F627}.Release|Any CPU.Build.0 = Release|Any CPU
-		{784EC1C5-0607-4B9B-A53A-B2636CE756A6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{784EC1C5-0607-4B9B-A53A-B2636CE756A6}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{784EC1C5-0607-4B9B-A53A-B2636CE756A6}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{784EC1C5-0607-4B9B-A53A-B2636CE756A6}.Release|Any CPU.Build.0 = Release|Any CPU
 		{F12B6A05-AB25-4D95-85FD-FA1BD7164FE2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{F12B6A05-AB25-4D95-85FD-FA1BD7164FE2}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{F12B6A05-AB25-4D95-85FD-FA1BD7164FE2}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -70,7 +64,6 @@ Global
 		{3A1422A3-AE3F-4631-A51E-07340367F48E} = {84114477-D195-4565-83C3-0F045FC52EA4}
 		{969F6A77-CCC1-4E81-91E5-F04B91994513} = {84114477-D195-4565-83C3-0F045FC52EA4}
 		{EF9F227B-52C2-4CF0-BB74-C5597BD5F627} = {84114477-D195-4565-83C3-0F045FC52EA4}
-		{784EC1C5-0607-4B9B-A53A-B2636CE756A6} = {84114477-D195-4565-83C3-0F045FC52EA4}
 		{F12B6A05-AB25-4D95-85FD-FA1BD7164FE2} = {84114477-D195-4565-83C3-0F045FC52EA4}
 		{7FF2C6BB-0DA3-4A3D-99DD-145B6ED7CF0D} = {0B38BBD6-2FA3-4AAE-960B-B264155C6802}
 		{C2215B36-B98E-4417-90CB-6808B00551C2} = {84114477-D195-4565-83C3-0F045FC52EA4}
