+++ src/Infrastructure.Persistence/Kathanika.Infrastructure.Persistence.csproj
@@ -5,6 +5,7 @@
 		<PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="7.0.0" />
 		<PackageReference Include="MongoDB.Driver" Version="2.21.0" />
 		<PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
+		<PackageReference Include="Microsoft.Extensions.Caching.Memory" Version="7.0.0" />
 	</ItemGroup>
 
 	<ItemGroup>
