+++ src/Web/Program.cs
@@ -1,7 +1,6 @@
 using Kathanika.Application;
 using Kathanika.Infrastructure.GraphQL;
 using Kathanika.Infrastructure.Persistence;
-using Kathanika.Infrastructure.Services;
 using Kathanika.Infrastructure.Workers;
 using Serilog;
 
@@ -19,7 +18,6 @@ try
     builder.Services.AddApplication()
         .AddGraphQLInfrastructure()
         .AddPersistenceInfrastructure(builder.Configuration)
-        .AddServicesInfrastructure(builder.Configuration)
         .AddWorkers(builder.Configuration);
 
     if (builder.Environment.IsDevelopment())
